feat: move iframe-resizer to addLibrariesJS
- getFirstExcerpt to load excerpt in JSON field via API
- parseHeaderBlog to load excerpt and image in JSON field via API
